backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
distro==1.8.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
openai==1.6.0
Pillow==10.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
zope.interface==6.1
